yeet

+++ Models/Customer.cs
@@ -14,6 +14,6 @@
 
         public string? Gender { get; set; }
 
-        public double? Age { get; set; }
+        public int? Age { get; set; }
     }
 }
